functions.c: removed some brackets on single line if in factorial multi.c/h, operator.c/h, sya.c, userFunctions.c: changed the increment type from int to uint16_t

+++ multi.c
@@ -493,7 +493,7 @@ matrix *extractValue(char **input, int varIndex, vari *var, error_return *error)
 }
 
 
-error_return checkVariable(const char *buffer, int *tok, char *input, int *iterator, vari *var, numberStack *num, operatorStack *ch) {
+error_return checkVariable(const char *buffer, int *tok, char *input, uint16_t *iterator, vari *var, numberStack *num, operatorStack *ch) {
 
 	error_return error = 0;
 
@@ -681,7 +681,7 @@ error_return printLine(char **input, vari *var) {
 }
 
 
-char **separateString(char *input, char const * const limiter, char const * const delimiter, int *iterator, error_return *error) {
+char **separateString(char *input, char const * const limiter, char const * const delimiter, uint16_t *iterator, error_return *error) {
 
 	//increment input to the first parenthesis
 	input += (*iterator + 1);
